modularizing some code and adding placeholder route components for further implementation once the backend is all the way set up
