--- app/javascript/components/pages/JiraIssues.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const JiraIssues = () => {
-  return (
-    <div className="text-center">Jira Issues</div>
-  )
-}
-
-export default JiraIssues
