+++ app/javascript/components/routes/Registration.js
