+++ app/javascript/components/routes/TicketShow.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const TicketShow = () => {
+  return (
+    <div>Ticket Show Page</div>
+  )
+}
+
+export default TicketShow
