+++ app/javascript/components/routes/Tickets.js
